Tweak generation kernel launching.
Since the number of work-items in generation kernel execution is arbitrary,
sometimes the auto defined workgroup size was 1. This made the computation
inefficient and slow (at least 5x slower than with a optimal workgroup size).
Now, the number of work-items is rounded to allow better workgroup sizes,
this could be seen as padding.